+++ otlet/__init__.py
@@ -28,5 +28,5 @@ from .types import *
 from .api import *
 from .exceptions import *
 
-__version__ = "0.6.0b1"
+__version__ = "0.6.0"
 __license__ = "MIT"