+++ otlet/cli/options.py
@@ -4,6 +4,7 @@ from .arguments import *
 
 if "releases" not in sys.argv:
     ARGUMENT_LIST["package"] = PACKAGE_ARGUMENT
+    ARGUMENT_LIST["package_version"] = PACKAGE_VERSION_ARGUMENT
 else:
     RELEASES_ARGUMENT_LIST["package"] = PACKAGE_ARGUMENT
 
@@ -17,16 +18,18 @@ class OtletArgumentParser(ArgumentParser):
         self.arguments = ARGUMENT_LIST
         self.releases_arguments = RELEASES_ARGUMENT_LIST
         for key, arg in self.arguments.items():
-            self.add_argument(*arg["opts"], **self.without_keys(arg, "opts"), dest=key)
-        if "releases" in sys.argv or "--help" in sys.argv:
-            self.subparsers = self.add_subparsers(parser_class=ArgumentParser)
+            self.add_argument(*arg["opts"], **self.without_keys(arg, ["opts"]), dest=key)
+        pro = [_ for _ in ["releases", "--help", "-h"] if _ in sys.argv] # i did this because i can
+        if pro:
+            self.subparsers = self.add_subparsers(parser_class=ArgumentParser, metavar="[ sub-commands ]")
             self.releases_subparser = self.subparsers.add_parser(
                 "releases",
                 description="List releases for a specified package", 
-                help="List releases for a specified package"
+                help="List releases for a specified package",
+                epilog="(c) 2022-present Noah Tanner, released under the terms of the MIT License",
             )
             for key,arg in self.releases_arguments.items():
-                self.releases_subparser.add_argument(*arg["opts"], **self.without_keys(arg, "opts"), dest=key)
+                self.releases_subparser.add_argument(*arg["opts"], **self.without_keys(arg, ["opts"]), dest=key)
 
     def without_keys(self, d, keys):
         return {x: d[x] for x in d if x not in keys}